fix: 템플 상세 api 디렉토리 수정
